Move HTML 4 translator to `writers._html4`
